Reorganiza ordem de atributos e construtor, altera tipo do ID passando de Integer para Long

+++ src/main/java/com/kleyton/gerenciador_de_pautas_api/repositories/AssociadoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.kleyton.gerenciador_de_pautas_api.models.Associado;
 
 @Repository
-public interface AssociadoRepository extends JpaRepository<Associado, Integer> {
+public interface AssociadoRepository extends JpaRepository<Associado, Long> {
 
 }
